@@ -109,17 +109,17 @@
     "pluralize": "8.0.0",
     "prop-types": "^15.7.2",
     "qs": "6.11.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
     "react-dnd": "15.1.2",
     "react-dnd-html5-backend": "15.1.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "3.1.4",
     "react-fast-compare": "^3.2.0",
     "react-helmet": "^6.1.0",
     "react-intl": "6.4.1",
-    "react-is": "^17.0.2",
-    "react-query": "3.24.3",
+    "react-is": "^18.2.0",
+    "react-query": "3.39.2",
     "react-redux": "8.0.5",
     "react-refresh": "0.14.0",
     "react-router-dom": "5.3.4",
@@ -141,14 +141,12 @@
     "yup": "^0.32.9"
   },
   "devDependencies": {
-    "@testing-library/dom": "8.19.0",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/dom": "8.20.0",
+    "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
     "duplicate-dependencies-webpack-plugin": "^1.0.2",
     "glob": "8.0.3",
     "msw": "1.2.1",
-    "react-test-renderer": "^17.0.2",
     "speed-measure-webpack-plugin": "1.5.0",
     "webpack-bundle-analyzer": "^4.8.0"
   },
