@@ -64,15 +64,14 @@
     "@storybook/react": "^6.5.16",
     "@strapi/design-system": "1.7.5",
     "@strapi/icons": "1.7.5",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
     "browserslist-to-esbuild": "1.2.0",
     "cross-env": "^7.0.3",
     "esbuild-loader": "^2.21.0",
     "history": "^4.9.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "5.3.4",
     "rimraf": "3.0.2",
     "styled-components": "5.3.3",
@@ -83,8 +82,8 @@
   "peerDependencies": {
     "@strapi/design-system": "1.7.5",
     "@strapi/icons": "1.7.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0",
     "react-router-dom": "^5.3.4",
     "styled-components": "^5.3.3"
   },
