@@ -47,29 +47,28 @@
     "react-dnd": "15.1.2",
     "react-helmet": "^6.1.0",
     "react-intl": "6.4.1",
-    "react-query": "3.24.3",
+    "react-query": "3.39.2",
     "react-redux": "8.0.5",
     "react-select": "5.7.0",
     "sharp": "0.32.0",
     "yup": "^0.32.9"
   },
   "devDependencies": {
-    "@testing-library/dom": "8.19.0",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "8.0.1",
+    "@testing-library/dom": "8.20.0",
+    "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
     "msw": "1.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "5.3.4",
-    "react-test-renderer": "^17.0.2",
+    "react-test-renderer": "^18.2.0",
     "styled-components": "5.3.3"
   },
   "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.4",
-    "styled-components": "^5.3.3"
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0",
+    "react-router-dom": "5.3.4",
+    "styled-components": "5.3.3"
   },
   "engines": {
     "node": ">=14.19.1 <=18.x.x",
