@@ -17,18 +17,18 @@
     "react-intl": "6.4.1"
   },
   "devDependencies": {
-    "@testing-library/react": "12.1.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@testing-library/react": "13.4.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "5.3.4",
     "styled-components": "5.3.3"
   },
   "peerDependencies": {
     "@strapi/strapi": "^4.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.4",
-    "styled-components": "^5.3.3"
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0",
+    "react-router-dom": "5.3.4",
+    "styled-components": "5.3.3"
   },
   "scripts": {
     "test:front": "run -T cross-env IS_EE=true jest --config ./jest.config.front.js",
