@@ -46,7 +46,7 @@
     "react-copy-to-clipboard": "^5.1.0",
     "react-helmet": "^6.1.0",
     "react-intl": "6.4.1",
-    "react-query": "3.24.3",
+    "react-query": "3.39.2",
     "react-redux": "8.0.5",
     "redux": "^4.2.1",
     "reselect": "^4.1.7",
@@ -56,18 +56,18 @@
   },
   "peerDependencies": {
     "@strapi/strapi": "^4.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.4",
-    "styled-components": "^5.3.3"
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0",
+    "react-router-dom": "5.3.4",
+    "styled-components": "5.3.3"
   },
   "devDependencies": {
     "@apidevtools/swagger-parser": "^10.1.0",
-    "@testing-library/react": "12.1.4",
+    "@testing-library/react": "13.4.0",
     "history": "^4.9.0",
     "msw": "1.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "5.3.4",
     "styled-components": "5.3.3"
   },
