@@ -699,7 +699,7 @@ describe('ADMIN | PAGES | AUTH | BaseLogin', () => {
                   >
                     <label
                       class="c7 c13 c14"
-                      for="1"
+                      for=":r0:"
                     >
                       Email
                       <span
@@ -716,7 +716,7 @@ describe('ADMIN | PAGES | AUTH | BaseLogin', () => {
                         aria-invalid="false"
                         aria-required="true"
                         class="c19"
-                        id="1"
+                        id=":r0:"
                         name="email"
                         placeholder="kai@doe.com"
                         value=""
@@ -734,7 +734,7 @@ describe('ADMIN | PAGES | AUTH | BaseLogin', () => {
                   >
                     <label
                       class="c7 c13 c14"
-                      for="3"
+                      for=":r2:"
                     >
                       Password
                       <span
@@ -751,7 +751,7 @@ describe('ADMIN | PAGES | AUTH | BaseLogin', () => {
                         aria-invalid="false"
                         aria-required="true"
                         class="c20 c21"
-                        id="3"
+                        id=":r2:"
                         name="password"
                         type="password"
                         value=""
@@ -803,7 +803,7 @@ describe('ADMIN | PAGES | AUTH | BaseLogin', () => {
                       <input
                         aria-label="rememberMe"
                         class="c30"
-                        id="5"
+                        id=":r4:"
                         name="rememberMe"
                         type="checkbox"
                       />
