@@ -612,7 +612,7 @@ describe('ADMIN | PAGES | AUTH | ResetPassword', () => {
           >
             <div>
               <button
-                aria-controls="0"
+                aria-controls=":r0:"
                 aria-disabled="false"
                 aria-expanded="false"
                 aria-haspopup="true"
@@ -700,7 +700,7 @@ describe('ADMIN | PAGES | AUTH | ResetPassword', () => {
                       >
                         <label
                           class="c6 c20 c21"
-                          for="2"
+                          for=":r1:"
                         >
                           Password
                           <span
@@ -713,12 +713,12 @@ describe('ADMIN | PAGES | AUTH | ResetPassword', () => {
                           class="c24 c25"
                         >
                           <input
-                            aria-describedby="2-hint"
+                            aria-describedby=":r1:-hint"
                             aria-disabled="false"
                             aria-invalid="false"
                             aria-required="true"
                             class="c26 c27"
-                            id="2"
+                            id=":r1:"
                             name="password"
                             type="password"
                             value=""
@@ -754,7 +754,7 @@ describe('ADMIN | PAGES | AUTH | ResetPassword', () => {
                         </div>
                         <p
                           class="c6 c34"
-                          id="2-hint"
+                          id=":r1:-hint"
                         >
                           Password must contain at least 8 characters, 1 uppercase, 1 lowercase and 1 number
                         </p>
@@ -770,7 +770,7 @@ describe('ADMIN | PAGES | AUTH | ResetPassword', () => {
                       >
                         <label
                           class="c6 c20 c21"
-                          for="4"
+                          for=":r3:"
                         >
                           Confirm Password
                           <span
@@ -787,7 +787,7 @@ describe('ADMIN | PAGES | AUTH | ResetPassword', () => {
                             aria-invalid="false"
                             aria-required="true"
                             class="c26 c27"
-                            id="4"
+                            id=":r3:"
                             name="confirmPassword"
                             type="password"
                             value=""
