@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import { IntlProvider } from 'react-intl';
 import { createMemoryHistory } from 'history';
 import { Router, Route } from 'react-router-dom';
@@ -31,7 +31,7 @@ describe('DynamicTable', () => {
   it('renders and matches the snapshot', () => {
     const history = createMemoryHistory();
     const pagination = { pageCount: 2 };
-    history.push('/settings/user?pageSize=10&page=1&sort=firstname');
+    act(() => history.push('/settings/user?pageSize=10&page=1&sort=firstname'));
     const app = makeApp(history, pagination);
 
     const { container } = render(app);
