@@ -642,7 +642,7 @@ describe('Admin | UseCasePage', () => {
             >
               <div>
                 <button
-                  aria-controls="0"
+                  aria-controls=":r0:"
                   aria-disabled="false"
                   aria-expanded="false"
                   aria-haspopup="true"
@@ -733,7 +733,7 @@ describe('Admin | UseCasePage', () => {
                         </label>
                         <div
                           aria-autocomplete="none"
-                          aria-controls="radix-0"
+                          aria-controls="radix-:r3:"
                           aria-describedby="usecase-hint usecase-error"
                           aria-expanded="false"
                           aria-label="What type of work do you do?"
