@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@testing-library/react';
 import axios from 'axios';
 
 import { useAppInfo } from '../AppInfo';
@@ -16,7 +16,7 @@ jest.mock('axios', () => ({
 
 const setup = (props) =>
   renderHook(() => useTracking(), {
-    wrapper: ({ children, ...restProps }) => (
+    wrapper: ({ children }) => (
       <TrackingProvider
         value={{
           uuid: '1',
@@ -24,13 +24,12 @@ const setup = (props) =>
             nestedProperty: true,
           },
           deviceId: 'someTestDeviceId',
-          ...restProps,
+          ...props,
         }}
       >
         {children}
       </TrackingProvider>
     ),
-    initialProps: props,
   });
 
 describe('useTracking', () => {
