@@ -1,20 +1,13 @@
-import React from 'react';
-import { render, act } from '@testing-library/react';
+import { renderHook } from '@testing-library/react';
 
 import useId from '../useId';
 
 function setup(...args) {
-  let returnVal;
+  const { result } = renderHook((...props) => useId(...props), {
+    initialProps: args,
+  });
 
-  function TestComponent() {
-    returnVal = useId(...args);
-
-    return null;
-  }
-
-  render(<TestComponent />);
-
-  return returnVal;
+  return result.current;
 }
 
 describe('useId', () => {
@@ -25,17 +18,13 @@ describe('useId', () => {
 
     expect(id).toBe('one-1');
 
-    act(() => {
-      id = setup('one');
-    });
+    id = setup('one');
 
     expect(id).toBe('one-2');
   });
 
   test('works with namespaces', () => {
-    act(() => {
-      id = setup('two');
-    });
+    id = setup('two');
 
     expect(id).toBe('two-3');
   });
