@@ -1,7 +1,7 @@
 import React from 'react';
 import { IntlProvider } from 'react-intl';
 import { QueryClientProvider, QueryClient, useQueryClient } from 'react-query';
-import { renderHook, act } from '@testing-library/react-hooks';
+import { renderHook, act, waitFor } from '@testing-library/react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import { NotificationsProvider, useNotification, useFetchClient } from '@strapi/helper-plugin';
@@ -176,7 +176,7 @@ describe('useBulkMove', () => {
     const toggleNotification = useNotification();
     const queryClient = useQueryClient();
 
-    const { result, waitFor } = await setup();
+    const { result } = await setup();
     const { move } = result.current;
 
     await act(async () => {
@@ -200,7 +200,6 @@ describe('useBulkMove', () => {
 
     const {
       result: { current },
-      waitFor,
     } = await setup();
     const { move } = current;
 
