@@ -1,7 +1,7 @@
 import React from 'react';
 import { IntlProvider } from 'react-intl';
 import { QueryClientProvider, QueryClient, useQueryClient } from 'react-query';
-import { renderHook, act } from '@testing-library/react-hooks';
+import { renderHook, act, waitFor } from '@testing-library/react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import { NotificationsProvider, useNotification, useFetchClient } from '@strapi/helper-plugin';
@@ -85,10 +85,10 @@ describe('useConfig', () => {
       const { get } = useFetchClient();
       get.mockReturnValueOnce(mockGetResponse);
 
-      const { waitFor, result } = await setup();
+      const { result } = await setup();
       expect(get).toHaveBeenCalledWith(`/${pluginId}/configuration`);
 
-      await waitFor(() => !result.current.config.isLoading);
+      await waitFor(() => expect(result.current.config.isLoading).toBe(false));
       expect(result.current.config.data).toEqual(mockGetResponse.data.data);
     });
 
@@ -100,11 +100,9 @@ describe('useConfig', () => {
         },
       });
 
-      const { waitFor, result } = await setup();
+      const { result } = await setup();
 
-      await waitFor(() => !result.current.config.isLoading);
-
-      expect(result.current.config.data).toEqual({});
+      await waitFor(() => expect(result.current.config.data).toEqual({}));
     });
 
     test('calls toggleNotification in case of error', async () => {
@@ -114,7 +112,7 @@ describe('useConfig', () => {
 
       get.mockRejectedValueOnce(new Error('Jest mock error'));
       const toggleNotification = useNotification();
-      const { waitFor } = await setup({});
+      await setup({});
 
       await waitFor(() =>
         expect(toggleNotification).toBeCalledWith({
