@@ -640,7 +640,7 @@ describe('Plugin | Documentation | PluginPage', () => {
                               <span>
                                 <button
                                   aria-disabled="false"
-                                  aria-labelledby="2"
+                                  aria-labelledby=":r2:"
                                   class="c33 c34 c12 c35"
                                   tabindex="-1"
                                   type="button"
@@ -675,7 +675,7 @@ describe('Plugin | Documentation | PluginPage', () => {
                               <span>
                                 <button
                                   aria-disabled="false"
-                                  aria-labelledby="3"
+                                  aria-labelledby=":r4:"
                                   class="c33 c34 c12 c35"
                                   tabindex="-1"
                                   type="button"
@@ -706,7 +706,7 @@ describe('Plugin | Documentation | PluginPage', () => {
                               <span>
                                 <button
                                   aria-disabled="false"
-                                  aria-labelledby="4"
+                                  aria-labelledby=":r6:"
                                   class="c33 c34 c12 c35"
                                   tabindex="-1"
                                   type="button"
@@ -774,7 +774,7 @@ describe('Plugin | Documentation | PluginPage', () => {
                               <span>
                                 <button
                                   aria-disabled="false"
-                                  aria-labelledby="5"
+                                  aria-labelledby=":r8:"
                                   class="c33 c34 c12 c35"
                                   tabindex="-1"
                                   type="button"
@@ -809,7 +809,7 @@ describe('Plugin | Documentation | PluginPage', () => {
                               <span>
                                 <button
                                   aria-disabled="false"
-                                  aria-labelledby="6"
+                                  aria-labelledby=":ra:"
                                   class="c33 c34 c12 c35"
                                   tabindex="-1"
                                   type="button"
