@@ -1,5 +1,5 @@
 import React from 'react';
-import { fireEvent, render, waitFor, waitForElementToBeRemoved } from '@testing-library/react';
+import { render, waitFor, act, waitForElementToBeRemoved } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { ThemeProvider, lightTheme } from '@strapi/design-system';
 import { Router, Switch, Route } from 'react-router-dom';
@@ -37,7 +37,7 @@ function makeAndRenderApp() {
     </IntlProvider>
   );
   const renderResult = render(app);
-  history.push(`/settings/${pluginId}/roles/1`);
+  act(() => history.push(`/settings/${pluginId}/roles/1`));
 
   return renderResult;
 }
@@ -879,9 +879,7 @@ describe('Admin | containers | RoleEditPage', () => {
           action="#"
           novalidate=""
         >
-          <div
-            style="height: 0px;"
-          >
+          <div>
             <div
               class="c1"
               data-strapi-header="true"
@@ -996,7 +994,7 @@ describe('Admin | containers | RoleEditPage', () => {
                             >
                               <label
                                 class="c9 c25 c26"
-                                for="1"
+                                for=":r0:"
                               >
                                 Name
                               </label>
@@ -1008,7 +1006,7 @@ describe('Admin | containers | RoleEditPage', () => {
                                   aria-invalid="false"
                                   aria-required="false"
                                   class="c28"
-                                  id="1"
+                                  id=":r0:"
                                   name="name"
                                   value="Authenticated"
                                 />
@@ -1038,7 +1036,7 @@ describe('Admin | containers | RoleEditPage', () => {
                               >
                                 <label
                                   class="c9 c25 c26"
-                                  for="3"
+                                  for=":r2:"
                                 >
                                   Description
                                 </label>
@@ -1050,7 +1048,7 @@ describe('Admin | containers | RoleEditPage', () => {
                                   aria-invalid="false"
                                   aria-required="false"
                                   class="c31"
-                                  id="3"
+                                  id=":r2:"
                                   name="description"
                                 >
                                   Default role given to authenticated user.
@@ -1105,10 +1103,10 @@ describe('Admin | containers | RoleEditPage', () => {
                               class="c5 c43 c44"
                             >
                               <button
-                                aria-controls="accordion-content-4"
+                                aria-controls="accordion-content-:r4:"
                                 aria-disabled="false"
                                 aria-expanded="false"
-                                aria-labelledby="accordion-label-4"
+                                aria-labelledby="accordion-label-:r4:"
                                 class="c5 c45 c12 c46 c47"
                                 data-strapi-accordion-toggle="true"
                                 type="button"
@@ -1118,13 +1116,13 @@ describe('Admin | containers | RoleEditPage', () => {
                                 >
                                   <span
                                     class="c9 c49"
-                                    id="accordion-label-4"
+                                    id="accordion-label-:r4:"
                                   >
                                     Address
                                   </span>
                                   <p
                                     class="c9 c38"
-                                    id="accordion-desc-4"
+                                    id="accordion-desc-:r4:"
                                   >
                                     Define all allowed actions for the api::address plugin.
                                   </p>
@@ -1226,7 +1224,7 @@ describe('Admin | containers | RoleEditPage', () => {
   it('can toggle the permissions accordions and actions', async () => {
     const user = userEvent.setup();
     // Create app and wait for loading
-    const { getByLabelText, queryByText, getByTestId, getByText, getAllByRole } =
+    const { getByLabelText, getByRole, queryByText, getByTestId, getByText, getAllByRole } =
       makeAndRenderApp();
     const loader = getByTestId('loader');
     await waitForElementToBeRemoved(loader);
@@ -1243,13 +1241,16 @@ describe('Admin | containers | RoleEditPage', () => {
     expect(getByText('POST')).toBeInTheDocument();
     expect(getByText('/addresses')).toBeInTheDocument();
 
-    // Select all actions with the "select all" checkbox
-    const [selectAllCheckbox, ...actionCheckboxes] = getAllByRole('checkbox');
-    expect(selectAllCheckbox.checked).toBe(false);
-    fireEvent.click(selectAllCheckbox);
-    actionCheckboxes.forEach((actionCheckbox) => {
-      expect(actionCheckbox.checked).toBe(true);
-    });
+    expect(getByRole('checkbox', { name: 'Select all' }).checked).toBe(false);
+
+    await user.click(getByRole('checkbox', { name: 'Select all' }));
+
+    getAllByRole('checkbox')
+      .slice(1, -1)
+      .forEach((actionCheckbox) => {
+        console.log(actionCheckbox.checked);
+        expect(actionCheckbox.checked).toBe(true);
+      });
 
     // Close the collapse
     await user.click(collapse);
