@@ -527,7 +527,7 @@ describe('Plugin | Users and Permissions | RoleListPage', () => {
                     <span>
                       <button
                         aria-disabled="false"
-                        aria-labelledby="0"
+                        aria-labelledby=":r0:"
                         class="c21 c22 c12 c23"
                         tabindex="0"
                         type="button"
